Moved popup logic
