+++ byu_pytest_utils/popup.py
@@ -1,12 +1,12 @@
-POPUP = False
+_POPUP = False
 
 
 def set_popup(value: bool):
     """Set whether popups should be used."""
-    global POPUP
+    global _POPUP
     POPUP = value
 
 
 def get_popup() -> bool:
     """Return the current POPUP setting."""
-    return POPUP
+    return _POPUP
